Initial commit
I am going to try to make a Gamelan Composing App. I've got a lot
of thinking to figure out how I am going to go about this, but I
think if I keep this simple it shouldn't be to difficult of an
application to make and that will give me another feather in my cap
or rather my first feather of note. My other two apps are hardly
worth mentioning.