Got notes.
I've got a basic system for notes in place. It doesn't handle 8th
notes or diacritical marking yet, and it really throws the UI for a
loop every time you open and close the selection menu, but it is a
start. I want to save now while it is working before I go onto the
next part of this.