Mostly back now.
Had to learn how react native was expecting things. Primarily I figured
that react native is not react. The mind set is the same but one is a
web frame work the other is a UI frame work. I now know that I can't use
any HTML tags, I just code in an HTML style.