It appears to be working.
I think I have succesfully transfered over to android studio proper.
Which makes this kind of my initial commit... again... I was really
uncertain at this point whether I would stick with React Native or not.
I have decided I will for two reasons. First: It is still React and as
such opperates in the React mind set. React is what I'm currently
interested in learning, so that is one point. The other reason is
actually iOS. I'm making something that I hope will get use by the
national music museum and other places like it. Odds are they use iPads
for everything so it would be quite beneficial to have the ability to
publish to the apple app store as well.