Moved over.
So appearently I never commited gatra.js. That's highly unfortunate, but
sounds like me. I need to get better at noticing untracked files. But
fortunately I had commited note.js and that was the one with most of the
work in it. So I copied those over and rebuilt everything. It deployed!
I'm going to get the display working again but it is nice to see all the
classes are in and running again.