Display back to functional.
I've got a ways to go to make it pretty, but if I can't get the file
manager aspects of this program to work then it will be all for naught.
I must be able to save and load files and I must be able to export to
PDF or none of this will be useful.