PDF authoring working!
I've got it! the base most proof of concept is working. I have proven
that I can do the three critical functions of this application: I can
save the songs to a file. I can read the saved songs in from a file. And
most importantly, I can export the song to a PDF. Everything after this
is making it look pretty. I jest, there is more than that, but it feels
like that is the primary thing that I need to do.