Started work on note selection modal.
I'm going to use a modal display to make the note choice really big and
easy so it's touch screen friendly. I started work on that but my
evening has run out of time. I got so far as opening to many modals for
my device to handle. Now I just need to figure out how to do it once.