Note selector basic layout in place.
I'm having trouble getting the two notes to be on the same row together.
I have learned to use the flexbox tonight. It certainly makes some of
these things easier but it isn't quite working the way I expect.
Troublesome. Anyway, I've got the note selector layout covering the
important things like two rows of notes and displaying what the note
currently is. Now I need to wire up the buttons and that should be this
part of it all set.