Mind bursting.
This app is currently at about the maximum size I can fit inside my head
at one time. Which is not to bad. I'm going to be breaking it into
components soon. What I have right now is effectively the editor
component of this application. But I've got it working. The note
component passes information all the way up to the app and it changes
the lines appropriately. Whew.