Quicky: Saving works.
I had to tweak a line to get it to work correctly, but saving now works.
It basically just worked immediately as soon as I finished my previous
effort. With the state variable being updated with everything bubbling
up from the lowest levels it was just as simple as save the date.