I hate it when I forget new files.
Added lines. I actually added this like 3 commits ago. Anyway. I decided
to use lines as part of the structure of the code since that was the
level that the editor would want to work with. This allows for the
potential of interlacing lines from multiple instruments. I'm not sure
if I'll permit that behavior in the editor, we shall see.