Document selection functional.
The list of documents is loaded from disk. This is great, just the
behavior I am looking for. Next week I can start a renewed effort on
making the editor. I should be able to just pull in a number of the
components I've made previously. Though this time I have decided to try
my hand at function components since they are the new hotness. They seem
to be an uglier way of accomplishing the same thing. Over all I am
pleased with this evening. From scratch to a functional menu is great.