Displaying is back.
I've got a basic display of Irama back. It's currently hard coded to get
around the constrains of a new document. This is just the basic
displaying of a document without any logic connected yet but I can
already tell my experience with the first go around is helping a lot.
I've managed to get everything to layout correctly using flex box this
time.