Now displaying multiple passages.
I've managed to get multiple passages to be displayed tonight. The song
I've been using, Asmaradana, has 3 passages, including the Buka. The
gatra grouping is not quite what I'd like but all 3 passages are now
displayed. I need to update the logic for saving to get it up to spec
with the new file structure.

Which brings me to: New file structure! I've decided on what I think is
a good compromise for the files. I'm doing a mix of what I did
previously, gatra's separated by tabs, lines seperated by \n, and mark
down. So it is roughtly:
line
line
line
line
line
line

That should be sufficently ridged and flexible to allow me to do what I
need and still save the files in a human readable format.