Had to rework the data structure.
It would seem react is not to found of having state hooks down in my
note components. so I have moved the note editing modal out of note and
will be putting it directly into the song screen. This does reduce the
amount of "HTML" being rendered which is a good thing. I have as a
result also had to rewrite the message passing and some of the layout
description. I may still get adding new lines in this evening. We'll
see.