Started on loading.
I started working on loading. But I think it needs to be handled outside
of the song screen. I just realized by putting it in the function
component it will load the file every time it re-renders. Not so good.
So I'll have to load the contents and pass it in. But I've got content
coming in and being parsed correctly, it's just a matter of getting it
to be assigned to the state object correctly.