State values were what was screwing me up.
So the note selector had this weird lagging issue. Its been bothering me
for weeks now but I've not been able to pin it down. I'm going to have
to do something similar for line editors so I had to figure it out. And
I did. The state value that I had put in that component was what was
causing the problem. I removed it so it is now stateless and depends
entirely on it's parent object and it works now. Thank goodness.

I fear performance issues. It takes nearly a second for everything to
propagate back and forth. This is acceptable for now, but I will
probably need to revisit this and find a way to improve performance.