Passage Label editing working.
That should be enough to do the most basic of operations. I can make a
new song, add passages, add lines to those passages. Change the name of
the song and of the passages. There is still a minor issue with saving
or loading where an extra space is getting tacked on to the front of the
passage name. After that if I can get a decent PDF export I should have
a minimum viable product.