Ready for PDFin.
I've enabled multidex support because my project is a bloatten.
Inevitable when you use third party libraries. But saves me time so
worth it. I also got the PDF library to import and author a PDF. So now
I just need to get it to author the content from my song into a PDF and
we have a minimum viable product.