Releasable.
I've gotten a alpha version up in the Playstore. In fiddling with it I
have realized there are a number of things that need to be ironed out,
but at least I've got something in the play store. Technically my 3 time
doing such. Can't say it is my third app published because it's not yet
published and I never ended up publishing cuddlefish. But I don't think
I'm far from an actual published app at this rate.