Unicode is a no go.
So I wanted to do this with Unicode and its combining characters. But it
would seem that such is not actually possible. The things that are kinda
sorta like the symbols used in gamelan are not actually close enough to
work for my needs. So I found a gamelan font and will be high jacking
the normal ascii character name space. this should be interesting.