KepatihanPro font works.
It is not pretty yet, but it works. I have diacritical markings working.
Appearently diacritical markings is not the most accurate way of
describing them. The whole notation system is called Kepatihan which is
where the font draws its name. I may change the name of this project
from Panerus to Kepatihan at some point in the future as that name makes
more sense.