Refactoring from the note.
I've begun a rather large refactor. This was brought about due to new
information that has come from the change from unicode to the kepatihan
font. Saddly my objective of having a plain save file won't be viable
any more since I am not using unicode anymore. Unicode was one thing but
g4 isn't going to make sense as a four with a cirlce around it. So I am
refactoring with notes as objects instead of strings. This will make a
lot of the logic around them much easier since I can address the parts
and the diacritical markings directly.

I've got a new note component completed and was working on updating
line. I ran aground for a while trying to get the layout of the line to
work since it isn't being forced from the bottom up anymore. Just got
it's width set in the right spot and now it is time for bed. But I am
glad to see some progress in a direction that should get me around my
current blockers.