Refactored bubbling.
I have decided to refactor the bubbling process for editing notes. I
have tried passing up each progressive tier of the structure as the
buble comes up but it was a little annoying to maintain. So the
alternative is to make the notes aware of where they are in the
structure. I have mixed feelings about this idea, but so far it appears
to be a cleaner implementation than the other way around. But the
cleaner implementation may simply be a better understanding of how it is
all pieced together on my part. We'll see.