Note selector starting to come back.
I'm starting to get the pieces to come back together. The note selector
is not currently making any changes but it does now pop up like it is
supposed to and display all the appropriate content. Huzzah! Now to get
it to edit the content and then I'll come back and get those diacritical
marking, the oh so important diacritical markings, back in.