Temporal delays
When you click to edit a note it doesn't select the note you just
selected it selects the previous note. I'm not sure why this is
happening but it has happened to me before. I'm going to call it an
evening here but note to me of next week: Look through the logs, find
where I solved this last time and adapt it to the new structure of
things.