Fixed saving and loading
Since I now use JSON for the structure of the song I can get away with
just saving the json in it's native form. That makes the loading and
saving of songs a lot less complicated, a welcome change. Not that it
was necessary to much as I did abstract the saving and loading, I could
still return to a plain text save option at some point in the future,
but for now I'll enjoy the simplicity of just using JSON.