Layouts and lines
Fixed the new line and new passage buttons. Everything seemed to be
working great but then scrolling wasn't. I figured out that the problem
was the box was to long and part of it was dropping off the bottom of
the display. This is something that could have taken me hours, weeks, to
solve but I managed to do it this evening between this commit and the
last. Man I have gotten better at this. So the trick was flexGrow. I had
to tell the button row to not grow but be a static size based on its
content and then the scroll view would fill the remaining space. Worked
perfectly, I was a little surprised.