New song was broken
Missed one of the other new buttons. It has a bunch of stuff that it
shouldn't need but does for some reason. I'll sort that out later, for
now I'm happy to have it working.

I have transcribed most of Asmaradana in the editor. Which is great, but
now, note for me of next week, a few diacritical markings that are
missing:

Downward swoop
Over head slash
Single Dot under
Single dot over

the dots over and under also need to be combined with the other
diacritical markings.