Diacritical marking buttons done
They aren't wired up yet, but I've got the octave buttons created. With
that I've got all the diacritical marking buttons made. So next week my
objective will be to wire the buttons up so that they can actually
change the value of the note. Which, that's an important note, the value
of the note, not of the diacritical marking, is what is changing. I also
need to do some work on the formating. The flex box distribution of the
note preview and the octave buttons is even. I'd really like to blow
up the font size of the preview. If I did that would take the extra
space that the preview is currently being given and make it make sense.