Just increased the font size
You know what, this looks pretty good like this. So I'm going to call
this good for now. I want to work on the document display of the home
screen next. If I can get that cleaned up a bit and double check that
PDF export is still working then I should be good to cut a beta version
and send to my tester.