A little nicer document view.
Visibly all I have done is setup the document so that it shows the title
provided in the editor rather than the path to the file. It's not a big
improvement visually but it does setup the underlying logic for doing
such which was the more interest part of this commit.