Started working on building.
I need to setup a process for making builds for this. I thought I got it
with these two commands but the playstore is complaining that I've not
signed it correctly. So not for next week: Get building working and
publish a new version and get pdf export working again.